Extract EmitterIA32Stat
